Rename src for package
