+++ montagu/__init__.py
@@ -1,0 +1,5 @@
+from montagu.montagu_api import MontaguAPI
+
+__all__ = [
+    MontaguAPI
+]
